requirements.txt
Flask
google-generativeai
markdown
